fix: move localstack to integration
