+++ internal/encryption/integration_test.go
@@ -3,15 +3,25 @@
 package encryption
 
 import (
+	"context"
 	"fmt"
+	"github.com/aws/aws-sdk-go-v2/aws"
 	"testing"
 	"time"
+
+	"github.com/TBD54566975/ftl/internal/log"
+	"github.com/TBD54566975/ftl/testutils"
 
 	"connectrpc.com/connect"
 	pbconsole "github.com/TBD54566975/ftl/backend/protos/xyz/block/ftl/v1/console"
 	in "github.com/TBD54566975/ftl/integration"
 	"github.com/TBD54566975/ftl/internal/slices"
 	"github.com/alecthomas/assert/v2"
+	"github.com/aws/aws-sdk-go-v2/service/kms"
+	awsv1 "github.com/aws/aws-sdk-go/aws"
+	awsv1credentials "github.com/aws/aws-sdk-go/aws/credentials"
+	awsv1session "github.com/aws/aws-sdk-go/aws/session"
+	awsv1kms "github.com/aws/aws-sdk-go/service/kms"
 )
 
 func TestEncryptionForLogs(t *testing.T) {
@@ -96,4 +106,41 @@ func validateAsyncCall(verb string, sensitive string) in.Action {
 		assert.Contains(t, string(request), "encrypted", "raw request string should not be stored in the table")
 		assert.NotContains(t, string(request), sensitive, "raw request string should not be stored in the table")
 	}
+}
+
+func TestKmsEncryptorLocalstack(t *testing.T) {
+	endpoint := "http://localhost:4566"
+
+	ctx := log.ContextWithNewDefaultLogger(context.Background())
+	cfg := testutils.NewLocalstackConfig(t, ctx)
+	v2client := kms.NewFromConfig(cfg, func(o *kms.Options) {
+		o.BaseEndpoint = aws.String(endpoint)
+	})
+	createKey, err := v2client.CreateKey(ctx, &kms.CreateKeyInput{})
+	assert.NoError(t, err)
+	uri := fmt.Sprintf("aws-kms://%s", *createKey.KeyMetadata.Arn)
+	fmt.Printf("URI: %s\n", uri)
+
+	// tink does not support awsv2 yet so here be dragons
+	// https://github.com/tink-crypto/tink-go-awskms/issues/2
+	s := awsv1session.Must(awsv1session.NewSession())
+	v1client := awsv1kms.New(s, &awsv1.Config{
+		Credentials: awsv1credentials.NewStaticCredentials("test", "test", ""),
+		Endpoint:    awsv1.String(endpoint),
+		Region:      awsv1.String("us-west-2"),
+	})
+
+	encryptor, err := NewKmsEncryptorGenerateKey(uri, v1client)
+	assert.NoError(t, err)
+
+	encrypted, err := encryptor.Encrypt(Logs, []byte("hunter2"))
+	assert.NoError(t, err)
+
+	decrypted, err := encryptor.Decrypt(Logs, encrypted)
+	assert.NoError(t, err)
+	assert.Equal(t, "hunter2", string(decrypted))
+
+	// Should fail to decrypt with the wrong subkey
+	_, err = encryptor.Decrypt(Async, encrypted)
+	assert.Error(t, err)
 }
